Fix delta time value with busy lopp and add sound manager with irrklang library (temp)

+++ Source/Core/Engine.h
@@ -30,8 +30,8 @@ namespace GameEngine {
 
 		float TargetTimePerFrame = (1.0f/60.0f);
 		int CurrentFPS=0;
-		float FrameTime = 0;
-		void Wait(float Milliseconds);
+		double FrameTime = 0;
+		void Wait(double Milliseconds);
 		GraphicContext* GContext;
 	};
